Master

+++ src/main.c
@@ -7,7 +7,7 @@ int main(void){
     FILE *fp;
     Calculator calculator = new_Calculator();
     boolean reply = true;
-    double a, b, result;
+    long double a, b, result;
     char op;
 
     setUpProtected(&calculator);
